requirements.txt
# Bot
aiogram==3.13.1

# API
fastapi==0.115.0
uvicorn[standard]==0.32.1
python-multipart==0.0.12
python-jose[cryptography]==3.3.0

# HTTP client (for adguard/federation)
httpx==0.28.0

# Database
sqlalchemy==2.0.36
aiosqlite==0.20.0

# Config / env
pydantic==2.10.1
pydantic-settings==2.6.1
python-dotenv==1.0.1

# Auth
bcrypt==4.1.3

# Templating (Nginx configs)
jinja2==3.1.4

# File handling
aiofiles==24.1.0

# QR codes
qrcode[pil]==8.0
Pillow==10.4.0
